// src/app/core/models/date-group.model.ts
import { DateGroupType } from "./date-group-type.model";

export interface DateGroup {
    id: number;              // bigint generated by default
    title: string | null;
    specific_date: string | null;
    day_of_week: number | null;
    month: number | null;
    day: number | null;
    hour: number | null;
    date_group_type_id: number | null;
    date_group_type : DateGroupType | null;
  }
  